Veikiantis /agmin/recipes

+++ finalproject1/src/main/java/lt/codeacademy/javau7/finalproject1/repositories/UserRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import lt.codeacademy.javau7.finalproject1.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByUsername(String username);
 
     Optional<User> findById(Long id);
 
-    void deleteById(Long id);}
+    void deleteById(int id);}
